Add `init_logger()` to `ab-cli-utils`
