+++ crates/shared/ab-cli-utils/src/lib.rs
@@ -2,6 +2,10 @@
 
 use std::panic;
 use std::process::exit;
+use tracing::level_filters::LevelFilter;
+use tracing_subscriber::layer::SubscriberExt;
+use tracing_subscriber::util::SubscriberInitExt;
+use tracing_subscriber::{EnvFilter, Layer, fmt};
 
 /// Install a panic handler which exits on panics, rather than unwinding. Unwinding can hang the
 /// tokio runtime waiting for stuck tasks or threads.
@@ -11,4 +15,17 @@ pub fn set_exit_on_panic() {
         default_panic_hook(panic_info);
         exit(1);
     }));
+}
+
+/// Initialize logger with typical settings
+pub fn init_logger() {
+    tracing_subscriber::registry()
+        .with(
+            fmt::layer().with_filter(
+                EnvFilter::builder()
+                    .with_default_directive(LevelFilter::INFO.into())
+                    .from_env_lossy(),
+            ),
+        )
+        .init();
 }
