+++ subspace/crates/subspace-farmer/Cargo.toml
@@ -16,6 +16,7 @@ name = "subspace-farmer"
 required-features = ["binary"]
 
 [dependencies]
+ab-cli-utils = { workspace = true, optional = true }
 ab-core-primitives = { workspace = true }
 ab-data-retrieval = { workspace = true }
 ab-direct-io-file = { workspace = true }
@@ -56,7 +57,6 @@ rayon = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 ss58-registry = { workspace = true }
-subspace-logging = { workspace = true, optional = true }
 subspace-metrics = { workspace = true, optional = true }
 subspace-networking = { workspace = true }
 subspace-proof-of-space-gpu = { workspace = true, optional = true }
@@ -89,10 +89,10 @@ _gpu = []
 #  while still being able to leverage default feature
 default-library = ["cluster", "numa"]
 binary = [
+    "dep:ab-cli-utils",
     "dep:clap",
     "dep:criterion",
     "dep:fdlimit",
     "dep:mimalloc",
     "dep:subspace-metrics",
-    "dep:subspace-logging",
 ]
