+++ subspace/crates/subspace-networking/Cargo.toml
@@ -16,6 +16,7 @@ include = [
 ]
 
 [dependencies]
+ab-cli-utils = { workspace = true }
 ab-core-primitives = { workspace = true, features = ["alloc", "scale-codec"] }
 async-lock = { workspace = true }
 async-trait = { workspace = true }
@@ -40,7 +41,6 @@ rand = { workspace = true }
 schnellru = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
-subspace-logging = { workspace = true }
 subspace-metrics = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["macros", "parking_lot", "rt-multi-thread", "signal", "sync", "time"] }
