+++ subspace/crates/subspace-networking/examples/networking.rs
@@ -1,5 +1,6 @@
 #![feature(type_changing_struct_update)]
 
+use ab_cli_utils::init_logger;
 use futures::StreamExt;
 use futures::channel::oneshot;
 use libp2p::gossipsub::Sha256Topic;
@@ -7,7 +8,6 @@ use libp2p::multiaddr::Protocol;
 use parking_lot::Mutex;
 use std::sync::Arc;
 use std::time::Duration;
-use subspace_logging::init_logger;
 use subspace_networking::Config;
 
 const TOPIC: &str = "Foo";
