+++ subspace/crates/subspace-node/src/commands.rs
@@ -1,7 +1,5 @@
 mod run;
-mod shared;
 mod wipe;
 
 pub use run::{RunOptions, run};
-pub(crate) use shared::set_exit_on_panic;
 pub use wipe::{WipeOptions, wipe};
