--- subspace/crates/subspace-node/src/commands/shared.rs
@@ -1,12 +1,0 @@
-use std::panic;
-use std::process::exit;
-
-/// Install a panic handler which exits on panics, rather than unwinding. Unwinding can hang the
-/// tokio runtime waiting for stuck tasks or threads.
-pub(crate) fn set_exit_on_panic() {
-    let default_panic_hook = panic::take_hook();
-    panic::set_hook(Box::new(move |panic_info| {
-        default_panic_hook(panic_info);
-        exit(1);
-    }));
-}
