--- subspace/shared/subspace-logging/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "subspace-logging"
-description = "Ensure all logging uilities are centralized for the whole project"
-license = "0BSD"
-version = "0.0.1"
-authors = ["Nazar Mokrynskyi <nazar@mokrynskyi.com>"]
-edition = "2024"
-include = [
-    "/src",
-    "/Cargo.toml",
-]
-
-[dependencies]
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true, features = ["env-filter"] }
-supports-color = { workspace = true }
