--- subspace/shared/subspace-logging/src/lib.rs
@@ -1,23 +1,0 @@
-use tracing::level_filters::LevelFilter;
-use tracing_subscriber::layer::SubscriberExt;
-use tracing_subscriber::util::SubscriberInitExt;
-use tracing_subscriber::{EnvFilter, Layer, fmt};
-
-pub fn init_logger() {
-    // TODO: Workaround for https://github.com/tokio-rs/tracing/issues/2214, also on
-    //  Windows terminal doesn't support the same colors as bash does
-    let enable_color = if cfg!(windows) {
-        false
-    } else {
-        supports_color::on(supports_color::Stream::Stderr).is_some()
-    };
-    tracing_subscriber::registry()
-        .with(
-            fmt::layer().with_ansi(enable_color).with_filter(
-                EnvFilter::builder()
-                    .with_default_directive(LevelFilter::INFO.into())
-                    .from_env_lossy(),
-            ),
-        )
-        .init();
-}
